+++ memberaudit_securegroups/models.py
@@ -19,6 +19,7 @@ from django.utils.translation import ngettext
 
 # Alliance Auth
 from allianceauth.authentication.models import CharacterOwnership
+from allianceauth.eveonline.models import EveCharacter
 
 # Member Audit
 from memberaudit.app_settings import MEMBERAUDIT_APP_NAME
@@ -212,9 +213,12 @@ class ActivityFilter(BaseFilter):
                     chars[char.user.pk].append(char.eve_character.character_name)
 
                 for char_user, char_list in chars.items():
-                    active_characters = ", ".join(char_list)
+                    message = ngettext(
+                        "Active Character: ", "Active Characters: ", len(char_list)
+                    )
+
                     output[char_user] = {
-                        "message": _(f"Active Characters: {active_characters}"),
+                        "message": message + ", ".join(sorted(char_list)),
                         "check": True,
                     }
 
@@ -286,7 +290,10 @@ class AgeFilter(BaseFilter):
                     chars[char.user.pk].append(char.eve_character.character_name)
 
                 for char_user, char_list in chars.items():
-                    output[char_user] = {"message": ", ".join(char_list), "check": True}
+                    output[char_user] = {
+                        "message": ", ".join(sorted(char_list)),
+                        "check": True,
+                    }
 
         return output
 
@@ -360,7 +367,7 @@ class AssetFilter(BaseFilter):
         output = defaultdict(lambda: {"message": "", "check": False})
 
         for character, char_list in chars.items():
-            output[character] = {"message": ", ".join(char_list), "check": True}
+            output[character] = {"message": ", ".join(sorted(char_list)), "check": True}
 
         return output
 
@@ -399,7 +406,14 @@ class ComplianceFilter(BaseFilter, SingletonModel):
         :rtype:
         """
 
-        output = defaultdict(lambda: {"message": "", "check": False})
+        output = defaultdict(
+            lambda: {
+                "message": _(
+                    f"Not all of your characters are added to {MEMBERAUDIT_APP_NAME}"
+                ),
+                "check": False,
+            }
+        )
 
         for user in users:
             if General.compliant_users().filter(pk=user.pk).exists():
@@ -408,6 +422,40 @@ class ComplianceFilter(BaseFilter, SingletonModel):
                         f"All characters have been added to {MEMBERAUDIT_APP_NAME}"
                     ),
                     "check": True,
+                }
+            else:
+                owned_chars_query = (
+                    EveCharacter.objects.filter(character_ownership__user=user)
+                    .select_related(
+                        "memberaudit_character",
+                        "memberaudit_character__location",
+                        "memberaudit_character__location__eve_solar_system",
+                        "memberaudit_character__location__eve_solar_system__eve_constellation__eve_region",
+                        "memberaudit_character__skillpoints",
+                        "memberaudit_character__unread_mail_count",
+                        "memberaudit_character__wallet_balance",
+                    )
+                    .order_by("character_name")
+                )
+
+                unregistered_chars = []
+
+                for eve_character in owned_chars_query:
+                    try:
+                        eve_character.memberaudit_character
+                    except AttributeError:
+                        unregistered_chars.append(eve_character.character_name)
+
+                missing_characters_message = ngettext(
+                    "Missing character: ",
+                    "Missing characters: ",
+                    len(unregistered_chars),
+                )
+
+                output[user.pk] = {
+                    "message": missing_characters_message
+                    + ", ".join(sorted(unregistered_chars)),
+                    "check": False,
                 }
 
         return output
@@ -476,7 +524,10 @@ class SkillPointFilter(BaseFilter):
                     chars[char.user.pk].append(char.eve_character.character_name)
 
                 for char_user, char_list in chars.items():
-                    output[char_user] = {"message": ", ".join(char_list), "check": True}
+                    output[char_user] = {
+                        "message": ", ".join(sorted(char_list)),
+                        "check": True,
+                    }
 
         return output
 
@@ -547,7 +598,7 @@ class SkillSetFilter(BaseFilter):
 
                         for char_user, char_list in chars.items():
                             output[char_user] = {
-                                "message": ", ".join(char_list),
+                                "message": ", ".join(sorted(char_list)),
                                 "check": True,
                             }