+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In Development]
 
+### Added
+
+- Names of missing characters when compliance filter fails
+
 ### Changed
 
 - Moved the build process to PEP 621 / pyproject.toml
+- Character names sorted alphabetically in all filters
 
 
 ## [0.3.0] - 2023-05-31